Add atomic backups
The Bedrock server can backup while running. For this it will need to be
put in hold, queried and resumed. The files retured in the query need to
be truncated to the given sizes for them to be valid.

This commit also contains a tar implementation that will process the
query result and return a tarball io stream.